requirements.txt
numpy==1.19.4
openai-clip==1.0.1
scipy==1.9.0
timm==1.0.7
tqdm==4.51.0
transformers==4.42.4
prettytable==2.0.0
gdown==3.12.2
imageio==2.9.0
sconf==0.2.3
tensorboard==2.3.0
Pillow==8.1.0
tensorboardX==2.1